requirements.txt
Django==1.9.2
django-filter==0.12.0
djangorestframework==3.3.2
Markdown==2.6.5
wheel==0.29.0
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.4.1
psycopg2==2.6.1
static3==0.6.1
Pillow==3.1.1
requests==2.9.1

